requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.0.0=pyhd8ed1ab_0
aiohttp=3.8.1=py37hcc03f2d_0
aiosignal=1.2.0=pyhd8ed1ab_0
ase=3.22.1=pypi_0
async-timeout=4.0.2=pyhd8ed1ab_0
asynctest=0.13.0=py_0
attrs=21.4.0=pyhd8ed1ab_0
blinker=1.4=py_1
boost=1.74.0=py37h457096e_5
boost-cpp=1.74.0=h54f0996_3
brotli=1.0.9=h8ffe710_6
brotli-bin=1.0.9=h8ffe710_6
brotlipy=0.7.0=py37hcc03f2d_1003
bzip2=1.0.8=h8ffe710_4
ca-certificates=2021.10.8=h5b45459_0
cached-property=1.5.2=pypi_0
cachetools=4.2.4=pyhd8ed1ab_0
cairo=1.16.0=h63a05c6_1001
certifi=2021.10.8=pypi_0
cffi=1.15.0=py37hd8e9650_0
charset-normalizer=2.0.11=pyhd8ed1ab_0
click=8.0.3=py37h03978a9_1
colorama=0.4.4=pyh9f0ad1d_0
cryptography=36.0.0=py37h21b164f_0
cycler=0.11.0=pyhd8ed1ab_0
dataclasses=0.8=pyhc8e2a94_3
fonttools=4.29.1=py37hcc03f2d_0
freetype=2.10.4=h546665d_1
frozenlist=1.3.0=py37hcc03f2d_0
google-auth=2.6.0=pyh6c4a22f_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
googledrivedownloader=0.4=pypi_0
greenlet=1.1.2=py37hf2a7229_1
grpcio=1.43.0=py37h265e445_0
h5py=3.6.0=pypi_0
icu=58.2=vc14hc45fdbb_0
idna=3.3=pyhd8ed1ab_0
importlib-metadata=4.10.1=py37h03978a9_0
intel-openmp=2022.0.0=h57928b3_3663
isodate=0.6.1=pypi_0
jinja2=3.0.3=pypi_0
joblib=1.1.0=pypi_0
jpeg=9b=vc14h4d7706e_1
kiwisolver=1.3.2=py37h8c56517_1
libblas=3.9.0=13_win64_mkl
libbrotlicommon=1.0.9=h8ffe710_6
libbrotlidec=1.0.9=h8ffe710_6
libbrotlienc=1.0.9=h8ffe710_6
libcblas=3.9.0=13_win64_mkl
liblapack=3.9.0=13_win64_mkl
libpng=1.6.37=h1d00b33_2
libprotobuf=3.19.4=h7755175_0
libtiff=4.2.0=hd0e1b90_0
libzlib=1.2.11=h8ffe710_1013
llvmlite=0.38.0=pypi_0
lz4-c=1.9.3=h8ffe710_1
markdown=3.3.6=pyhd8ed1ab_0
markupsafe=2.0.1=pypi_0
matplotlib-base=3.5.1=py37h4a79c79_0
mkl=2022.0.0=h0e2418a_796
multidict=6.0.2=py37hcc03f2d_0
munkres=1.1.4=pyh9f0ad1d_0
networkx=2.6.3=pypi_0
numba=0.55.1=pypi_0
numpy=1.21.5=py37h5fa1a60_0
oauthlib=3.2.0=pyhd8ed1ab_0
olefile=0.46=pyh9f0ad1d_1
openssl=3.0.0=h8ffe710_2
packaging=21.3=pyhd8ed1ab_0
pandas=1.3.5=pypi_0
pillow=9.0.1=pypi_0
pip=22.0.3=pyhd8ed1ab_0
pixman=0.38.0=hfa6e2cd_1003
protobuf=3.19.4=py37hf2a7229_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycairo=1.20.1=py37h43013c1_1
pycparser=2.21=pyhd8ed1ab_0
pyjwt=2.3.0=pyhd8ed1ab_1
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.7=pyhd8ed1ab_0
pysocks=1.7.1=py37h03978a9_4
python=3.7.12=h900ac77_100_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-louvain=0.16=pypi_0
python_abi=3.7=2_cp37m
pytz=2021.3=pyhd8ed1ab_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0=pypi_0
rdflib=6.1.1=pypi_0
rdkit=2021.09.4=py37hd1be60c_0
reportlab=3.5.68=py37h375b626_1
requests=2.27.1=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.8=pyhd8ed1ab_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
setuptools=59.5.0=pypi_0
six=1.16.0=pyh6c4a22f_0
sqlalchemy=1.4.31=py37hcc03f2d_0
sqlite=3.37.0=h8ffe710_0
tbb=2021.5.0=h2d74725_0
tensorboard=2.8.0=pyhd8ed1ab_0
tensorboard-data-server=0.6.0=py37h03978a9_1
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.11=h8ffe710_1
torch=1.8.0=pypi_0
torch-geometric=1.6.3=pypi_0
torch-scatter=2.0.6=pypi_0
torch-sparse=0.6.9=pypi_0
torchvision=0.8.2+cu110=pypi_0
tqdm=4.62.3=pypi_0
typing-extensions=4.0.1=hd8ed1ab_0
typing_extensions=4.0.1=pyha770c72_0
ucrt=10.0.20348.0=h57928b3_0
unicodedata2=14.0.0=py37hcc03f2d_0
urllib3=1.26.8=pyhd8ed1ab_1
vc=14.2=hb210afc_6
vs2015_runtime=14.29.30037=h902a5da_6
werkzeug=2.0.2=pyhd8ed1ab_0
wheel=0.37.1=pyhd8ed1ab_0
win_inet_pton=1.1.0=py37h03978a9_3
xz=5.2.5=h62dcd97_1
yarl=1.7.2=py37hcc03f2d_1
zipp=3.7.0=pyhd8ed1ab_1
zlib=1.2.11=h8ffe710_1013
zstd=1.4.9=h6255e5f_0
